+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.8.0] - 2021-11-20
 ### Added
 - Improved build pipeline to use [codaamok.build](https://github.com/codaamok/codaamok.build)
 - New `-ForwardQuery` boolean parameter in `New-ShlinkUrl` and `Set-ShlinkUrl` (new in Shlink 2.9.0).
@@ -86,7 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/codaamok/PSShlink/compare/0.7.0..HEAD
+[Unreleased]: https://github.com/codaamok/PSShlink/compare/0.8.0..HEAD
+[0.8.0]: https://github.com/codaamok/PSShlink/compare/0.7.0..0.8.0
 [0.7.0]: https://github.com/codaamok/PSShlink/compare/0.6.0..0.7.0
 [0.6.0]: https://github.com/codaamok/PSShlink/compare/0.5.0..0.6.0
 [0.5.0]: https://github.com/codaamok/PSShlink/compare/0.4.0..0.5.0